+++ src/main/java/tetrago/pyrros/common/block/ModBlocks.java
@@ -20,6 +20,7 @@ public class ModBlocks
     public static final RegistryObject<Block> ENCASED_MACHINE_FRAME = BLOCKS.register("encased_machine_frame", () -> new MultiblockComponentBlock(BlockBehaviour.Properties.of(Material.STONE).strength(1.5f, 6)));
     public static final RegistryObject<Block> REINFORCED_MACHINE_FRAME = BLOCKS.register("reinforced_machine_frame", () -> new MultiblockComponentBlock(BlockBehaviour.Properties.copy(ENCASED_MACHINE_FRAME.get())));
     public static final RegistryObject<Block> ROLLER = BLOCKS.register("roller", RollerBlock::new);
+    public static final RegistryObject<Block> BEARING = BLOCKS.register("bearing", () -> new FramedBlock(BlockBehaviour.Properties.of(Material.METAL).strength(1.5f, 6)));
     public static final RegistryObject<Block> BASIC_MOTOR = BLOCKS.register("basic_motor", () -> new MotorBlock(100));
     public static final RegistryObject<Block> STEEL_BLOCK = BLOCKS.register("steel_block", () -> new Block(BlockBehaviour.Properties.copy(Blocks.IRON_BLOCK)));
     public static final RegistryObject<Block> TITANIUM_BLOCK = BLOCKS.register("titanium_block", () -> new Block(BlockBehaviour.Properties.copy(Blocks.IRON_BLOCK)));
