+++ src/main/resources/assets/pyrros/models/block/bearing.json
@@ -1,0 +1,132 @@
+{
+	"parent": "minecraft:block/cube",
+	"textures": {
+		"0": "pyrros:block/basic_motor",
+		"1": "pyrros:block/motor_top",
+		"2": "pyrros:block/machine_frame",
+		"particle": "#0"
+	},
+	"elements": [
+		{
+			"from": [5, 11, 0],
+			"to": [11, 16, 16],
+			"faces": {
+				"north": {"uv": [0, 0, 6, 5], "texture": "#0"},
+				"east": {"uv": [0, 0, 16, 5], "texture": "#0"},
+				"south": {"uv": [0, 0, 6, 5], "texture": "#0"},
+				"west": {"uv": [0, 0, 16, 5], "texture": "#0"},
+				"up": {"uv": [5, 0, 11, 16], "texture": "#1"},
+				"down": {"uv": [5, 0, 11, 16], "texture": "#1"}
+			}
+		},
+		{
+			"from": [3, 11, 1],
+			"to": [5, 16, 15],
+			"faces": {
+				"north": {"uv": [0, 0, 2, 5], "texture": "#0"},
+				"east": {"uv": [0, 0, 14, 5], "texture": "#missing"},
+				"south": {"uv": [0, 0, 2, 5], "texture": "#0"},
+				"west": {"uv": [0, 0, 14, 5], "texture": "#0"},
+				"up": {"uv": [3, 1, 5, 15], "texture": "#1"},
+				"down": {"uv": [3, 1, 5, 15], "texture": "#1"}
+			}
+		},
+		{
+			"from": [11, 11, 1],
+			"to": [13, 16, 15],
+			"faces": {
+				"north": {"uv": [0, 0, 2, 5], "texture": "#0"},
+				"east": {"uv": [0, 0, 14, 5], "texture": "#0"},
+				"south": {"uv": [0, 0, 2, 5], "texture": "#0"},
+				"west": {"uv": [0, 0, 14, 5], "texture": "#missing"},
+				"up": {"uv": [11, 1, 13, 15], "texture": "#1"},
+				"down": {"uv": [11, 1, 13, 15], "texture": "#1"}
+			}
+		},
+		{
+			"from": [2, 11, 2],
+			"to": [3, 16, 14],
+			"faces": {
+				"north": {"uv": [0, 0, 1, 5], "texture": "#0"},
+				"east": {"uv": [0, 0, 12, 5], "texture": "#missing"},
+				"south": {"uv": [0, 0, 1, 5], "texture": "#0"},
+				"west": {"uv": [0, 0, 12, 5], "texture": "#0"},
+				"up": {"uv": [2, 2, 3, 14], "texture": "#1"},
+				"down": {"uv": [2, 2, 3, 14], "texture": "#1"}
+			}
+		},
+		{
+			"from": [13, 11, 2],
+			"to": [14, 16, 14],
+			"faces": {
+				"north": {"uv": [0, 0, 1, 5], "texture": "#0"},
+				"east": {"uv": [0, 0, 12, 5], "texture": "#0"},
+				"south": {"uv": [0, 0, 1, 5], "texture": "#0"},
+				"west": {"uv": [0, 0, 12, 5], "texture": "#missing"},
+				"up": {"uv": [13, 2, 14, 14], "texture": "#1"},
+				"down": {"uv": [13, 2, 14, 14], "texture": "#1"}
+			}
+		},
+		{
+			"from": [1, 11, 3],
+			"to": [2, 16, 13],
+			"faces": {
+				"north": {"uv": [0, 0, 1, 5], "texture": "#0"},
+				"east": {"uv": [0, 0, 10, 5], "texture": "#missing"},
+				"south": {"uv": [0, 0, 1, 5], "texture": "#0"},
+				"west": {"uv": [0, 0, 10, 5], "texture": "#0"},
+				"up": {"uv": [1, 3, 2, 13], "texture": "#1"},
+				"down": {"uv": [1, 3, 2, 13], "texture": "#1"}
+			}
+		},
+		{
+			"from": [14, 11, 3],
+			"to": [15, 16, 13],
+			"faces": {
+				"north": {"uv": [0, 0, 1, 5], "texture": "#0"},
+				"east": {"uv": [0, 0, 10, 5], "texture": "#0"},
+				"south": {"uv": [0, 0, 1, 5], "texture": "#0"},
+				"west": {"uv": [0, 0, 10, 5], "texture": "#missing"},
+				"up": {"uv": [14, 3, 15, 13], "texture": "#1"},
+				"down": {"uv": [14, 3, 15, 13], "texture": "#1"}
+			}
+		},
+		{
+			"from": [0, 11, 5],
+			"to": [1, 16, 11],
+			"faces": {
+				"north": {"uv": [0, 0, 1, 5], "texture": "#0"},
+				"east": {"uv": [0, 0, 6, 5], "texture": "#missing"},
+				"south": {"uv": [0, 0, 1, 5], "texture": "#0"},
+				"west": {"uv": [0, 0, 6, 5], "texture": "#0"},
+				"up": {"uv": [0, 5, 1, 11], "texture": "#1"},
+				"down": {"uv": [0, 5, 1, 11], "texture": "#1"}
+			}
+		},
+		{
+			"from": [15, 11, 5],
+			"to": [16, 16, 11],
+			"faces": {
+				"north": {"uv": [0, 0, 1, 5], "texture": "#0"},
+				"east": {"uv": [0, 0, 6, 5], "texture": "#0"},
+				"south": {"uv": [0, 0, 1, 5], "texture": "#0"},
+				"west": {"uv": [0, 0, 6, 5], "texture": "#missing"},
+				"up": {"uv": [15, 5, 16, 11], "texture": "#1"},
+				"down": {"uv": [15, 5, 16, 11], "texture": "#1"}
+			}
+		},
+		{
+			"from": [-0.1, -0.1, -0.1],
+			"to": [16.1, 16.1, 16.1],
+			"faces": {
+				"north": {"uv": [0, 0, 16, 16], "texture": "#2"},
+				"east": {"uv": [0, 0, 16, 16], "texture": "#2"},
+				"south": {"uv": [0, 0, 16, 16], "texture": "#2"},
+				"west": {"uv": [0, 0, 16, 16], "texture": "#2"},
+				"up": {"uv": [0, 0, 16, 16], "texture": "#2"},
+				"down": {"uv": [0, 0, 16, 16], "texture": "#2"}
+			}
+		}
+	],
+	"display": {}
+}
